backend/pyproject.toml
[tool.poetry]
name = "twitter"
version = "0.1.0"
description = ""
authors = ["Matthew Coyle <mattcoyle48@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.102.0"
python-jose = "^3.3.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
coverage = "^7.3.0"
pytest-bdd = "^6.1.1"


[tool.poetry.group.dev.dependencies]
invoke = "^2.2.0"
black = "^23.7.0"
isort = "^5.12.0"
pylint = "^2.17.5"
mypy = "^1.5.1"
types-python-jose = "^3.3.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
disable = "C0114, C0115, C0116, R0903"

[tool.isort]
profile = "black"